requirements.txt
ansible==1.9.4
pyyaml
jinja2
